--- flint-extras/src/nmod_poly_mat_extra/nmod_poly_mat_middle_product_naive.c
@@ -1,26 +1,0 @@
-/*
-    Copyright (C) 2025 Éric Schost
-
-    This file is part of PML.
-
-    PML is free software: you can redistribute it and/or modify it under
-    the terms of the GNU General Public License version 2.0 (GPL-2.0-or-later)
-    as published by the Free Software Foundation; either version 2 of the
-    License, or (at your option) any later version. See
-    <https://www.gnu.org/licenses/>.
-*/
-
-#include <flint/nmod_poly_mat.h>
-
-/** Middle product for polynomial matrices
- *  sets C = ((A * B) div x^dA) mod x^(dB+1), assuming deg(A) <= dA and deg(B) <= dA + dB
- *  output can alias input
- *  naive implementation (multiply, shift, truncate)
- */
-void nmod_poly_mat_middle_product_naive(nmod_poly_mat_t C, const nmod_poly_mat_t A, const nmod_poly_mat_t B,
-                                        const ulong dA, const ulong dB)
-{
-    nmod_poly_mat_mul(C, A, B);
-    nmod_poly_mat_shift_right(C, C, dA);
-    nmod_poly_mat_truncate(C, dB + 1);
-}
